requirements.txt
streamlit
openai
google-generativeai
gspread
google-auth-oauthlib
googlesearch-python
pandas
python-dotenv
pillow